move useAuth hook
